# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SETU-DRR API",
    "description": "Hazard Red Zone Identification, Carrying Capacity Assessment & Relocation Decision Support Platform.",
    "version": "v1.0.0"
  },
  "paths": {
...
    "/zones": {
      "get": {
        "tags": [
          "Hazard Zones"
        ],
        "summary": "Query H3 hazard zones within spatial viewport",
        "description": "Retrieves active Multi-Hazard Index and dominant hazard classifications for H3 hexagonal grid cells within the specified bounding box.",
        "operationId": "get_zones_zones_get",
        "parameters": [
          {
            "name": "bbox",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Bounding box as 'min_lon,min_lat,max_lon,max_lat' (e.g. '75.8,11.5,76.3,11.9')",
              "examples": [
                "75.8,11.5,76.3,11.9"
              ],
              "title": "Bbox"
            },
            "description": "Bounding box as 'min_lon,min_lat,max_lon,max_lat' (e.g. '75.8,11.5,76.3,11.9')"
          },
          {
            "name": "res",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 9,
              "minimum": 6,
              "description": "H3 grid resolution (6=Overview, 7=National, 8=Pilot, 9=Site)",
              "default": 8,
              "title": "Res"
            },
            "description": "H3 grid resolution (6=Overview, 7=National, 8=Pilot, 9=Site)"
          },
          {
            "name": "valid_at",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Historical timestamp to evaluate (ISO 8601). Defaults to latest.",
              "title": "Valid At"
            },
            "description": "Historical timestamp to evaluate (ISO 8601). Defaults to latest."
          },
          {
            "name": "admin",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by Administrative Unit ID or LGD Code (e.g. 555 for Wayanad)",
              "title": "Admin"
            },
            "description": "Filter by Administrative Unit ID or LGD Code (e.g. 555 for Wayanad)"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 5000,
              "minimum": 1,
              "description": "Maximum number of cells to return (max 5000).",
              "default": 1000,
              "title": "Limit"
            },
            "description": "Maximum number of cells to return (max 5000)."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ZoneCellSummary"
                  },
                  "title": "Response Get Zones Zones Get"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Invalid parameters or malformed input format.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Not Found - Requested resource, cell, or entity does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error - Request parameter or payload validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - An unexpected system or database error occurred.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable - No valid serving version is active. Pipeline data is not ready.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/zones/{h3}": {
      "get": {
        "tags": [
          "Hazard Zones"
        ],
        "summary": "Get full cell detail and SHAP/heuristic explanation",
        "description": "Retrieves multi-hazard breakdown, static/live MHI, and feature attributions for a single H3 cell.",
        "operationId": "get_zone_detail_zones__h3__get",
        "parameters": [
          {
            "name": "h3",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "H3 index as hexadecimal string (e.g. '8860064989fffff')",
              "examples": [
                "8860064989fffff"
              ],
              "title": "H3"
            },
            "description": "H3 index as hexadecimal string (e.g. '8860064989fffff')"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ZoneCellDetail"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Invalid parameters or malformed input format.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Not Found - Requested resource, cell, or entity does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error - Request parameter or payload validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - An unexpected system or database error occurred.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable - No valid serving version is active. Pipeline data is not ready.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
    "/sites/{id}/capacity": {
      "post": {
        "tags": [
          "Candidate Sites & Capacity"
        ],
        "summary": "Recompute candidate site carrying capacity with overridden policy norms",
        "description": "Simulates carrying capacity under modified policy parameters (e.g. plot area, LPCD, spare school/health capacity). Returns the baseline capacity, scenario capacity, net delta in supportable households, and augmented relief options. Requires authenticated user with 'capacity.recompute' permission (Government Official) and authorized jurisdiction scope.",
        "operationId": "recompute_site_capacity_sites__id__capacity_post",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "Candidate Site ID (integer primary key).",
              "examples": [
                1
              ],
              "title": "Id"
            },
            "description": "Candidate Site ID (integer primary key)."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SiteCapacityOverrideRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteCapacityOverrideResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated - Missing, invalid, expired, or revoked session cookie.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Insufficient permissions or role.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Not Found - Requested resource, cell, or entity does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error - Request parameter or payload validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - An unexpected system or database error occurred.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable - No valid serving version is active. Pipeline data is not ready.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/sites/{id}": {
      "get": {
        "tags": [
          "Candidate Sites & Capacity"
        ],
        "summary": "Get candidate site detail by ID",
        "description": "Retrieves full candidate relocation site profile including GeoJSON polygon geometry and resource capacity breakdown.",
        "operationId": "get_site_detail_sites__id__get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "Candidate Site ID (integer primary key).",
              "examples": [
                1
              ],
              "title": "Id"
            },
            "description": "Candidate Site ID (integer primary key)."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateSiteDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not Found - Requested resource, cell, or entity does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error - Request parameter or payload validation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - An unexpected system or database error occurred.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable - No valid serving version is active. Pipeline data is not ready.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Standardized machine-readable error code string."
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable explanation of the error."
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id",
            "description": "Unique request tracing identifier."
          },
          "details": {
            "additionalProperties": true,
            "type": "object",
            "title": "Details",
            "description": "Contextual error details or validation breakdown."
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail",
        "description": "Standardized error details conforming to SETU-DRR API error envelope."
      },
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail",
            "description": "Error details payload."
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "description": "Canonical API error response envelope."
      },
...
      "SiteCapacityOverrideRequest": {
        "properties": {
          "plot_area_m2": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1000.0,
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Plot Area M2",
            "description": "Override plot area per HH."
          },
          "water_lpcd": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 500.0,
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Water Lpcd",
            "description": "Override LPCD norm (55 rural, 135 urban)."
          },
          "daily_water_yield_liters": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Daily Water Yield Liters",
            "description": "Override sustainable water yield."
          },
          "spare_school_seats": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Spare School Seats",
            "description": "Override spare school seats."
          },
          "spare_health_capacity_pop": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Spare Health Capacity Pop",
            "description": "Override spare PHC population."
          },
          "livelihood_multiplier": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Livelihood Multiplier",
            "description": "Override livelihood multiplier."
          }
        },
        "type": "object",
        "title": "SiteCapacityOverrideRequest",
        "description": "Request payload for capacity scenario simulation (POST /sites/{id}/capacity)."
      },
      "SiteCapacityOverrideResponse": {
        "properties": {
          "site_id": {
            "type": "integer",
            "title": "Site Id"
          },
          "base_capacity": {
            "$ref": "#/components/schemas/CapacityBreakdownDTO"
          },
          "scenario_capacity": {
            "$ref": "#/components/schemas/CapacityBreakdownDTO"
          },
          "delta_households": {
            "type": "integer",
            "title": "Delta Households",
            "description": "Change in final household capacity under override scenario."
          },
          "augmented_options": {
            "items": {
              "$ref": "#/components/schemas/AugmentedCapacityDTO"
            },
            "type": "array",
            "title": "Augmented Options"
          },
          "screening_grade": {
            "type": "string",
            "title": "Screening Grade",
            "description": "Persistent decision-support disclaimer notice.",
            "default": "Screening Grade: Cell-level screening and prioritisation tool. Geotechnical investigation, hydraulic study, and community consultation required before executing relocation orders."
          }
        },
        "type": "object",
        "required": [
          "site_id",
          "base_capacity",
          "scenario_capacity",
          "delta_households"
        ],
        "title": "SiteCapacityOverrideResponse",
        "description": "Response payload for capacity scenario simulation (POST /sites/{id}/capacity)."
      },
...
      "ZoneCellDetail": {
        "properties": {
          "h3": {
            "type": "string",
            "title": "H3"
          },
          "h3_int": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "H3 Int"
          },
          "res": {
            "type": "integer",
            "title": "Res"
          },
          "dataset_version": {
            "type": "string",
            "title": "Dataset Version",
            "default": "demo-day2-v1"
          },
          "model_version": {
            "type": "string",
            "title": "Model Version",
            "default": "baseline-v1"
          },
          "data_quality": {
            "type": "string",
            "title": "Data Quality",
            "default": "synthetic"
          },
          "valid_at": {
            "type": "string",
            "format": "date-time",
            "title": "Valid At"
          },
          "admin_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Admin Id"
          },
          "admin_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Admin Name"
          },
          "habitation_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Habitation Id"
          },
          "habitation_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Habitation Name"
          },
          "population": {
            "type": "number",
            "title": "Population",
            "default": 0.0
          },
          "built_area_m2": {
            "type": "number",
            "title": "Built Area M2",
            "default": 0.0
          },
          "centroid": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Centroid"
          },
          "mhi_static": {
            "type": "number",
            "title": "Mhi Static"
          },
          "mhi_live": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mhi Live"
          },
          "mhi_fcst": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mhi Fcst"
          },
          "dominant_hazard": {
            "type": "string",
            "title": "Dominant Hazard"
          },
          "zone_class": {
            "$ref": "#/components/schemas/ZoneClass"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence",
            "default": 1.0
          },
          "hazards": {
            "items": {
              "$ref": "#/components/schemas/HazardDetailDTO"
            },
            "type": "array",
            "title": "Hazards"
          },
          "explanation": {
            "items": {
              "$ref": "#/components/schemas/FeatureContributionDTO"
            },
            "type": "array",
            "title": "Explanation"
          },
          "screening_grade": {
            "type": "string",
            "title": "Screening Grade",
            "default": "Screening Grade: Cell-level screening and prioritisation tool. Geotechnical investigation, hydraulic study, and community consultation required before executing relocation orders."
          }
        },
        "type": "object",
        "required": [
          "h3",
          "res",
          "valid_at",
          "centroid",
          "mhi_static",
          "dominant_hazard",
          "zone_class"
        ],
        "title": "ZoneCellDetail",
        "description": "Full cell detail with SHAP/heuristic explanation (GET /zones/{h3})."
      },
      "ZoneCellSummary": {
        "properties": {
          "h3": {
            "type": "string",
            "title": "H3",
            "description": "H3 index as hexadecimal string."
          },
          "h3_int": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "H3 Int",
            "description": "H3 index as 64-bit integer."
          },
          "res": {
            "type": "integer",
            "title": "Res",
            "description": "H3 resolution (6, 7, 8, 9)."
          },
          "mhi": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Mhi",
            "description": "Active Multi-Hazard Index in [0, 1]."
          },
          "mhi_static": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Mhi Static",
            "description": "Static Multi-Hazard Index in [0, 1]."
          },
          "mhi_live": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Mhi Live"
          },
          "mhi_fcst": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Mhi Fcst"
          },
          "dominant_hazard": {
            "type": "string",
            "title": "Dominant Hazard"
          },
          "zone_class": {
            "$ref": "#/components/schemas/ZoneClass"
          },
          "dataset_version": {
            "type": "string",
            "title": "Dataset Version",
            "description": "Published dataset version.",
            "default": "demo-day2-v1"
          },
          "model_version": {
            "type": "string",
            "title": "Model Version",
            "description": "ML model version.",
            "default": "baseline-v1"
          },
          "data_quality": {
            "type": "string",
            "title": "Data Quality",
            "description": "Provenance / screening grade quality.",
            "default": "synthetic"
          },
          "population": {
            "type": "number",
            "minimum": 0.0,
            "title": "Population",
            "default": 0.0
          },
          "built_area_m2": {
            "type": "number",
            "minimum": 0.0,
            "title": "Built Area M2",
            "default": 0.0
          },
          "centroid": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Centroid",
            "description": "[longitude, latitude]"
          }
        },
        "type": "object",
        "required": [
          "h3",
          "res",
          "mhi",
          "mhi_static",
          "dominant_hazard",
          "zone_class",
          "centroid"
        ],
        "title": "ZoneCellSummary",
        "description": "Summary record returned in spatial query list (GET /zones)."
      },
      "ZoneClass": {
        "type": "string",
        "enum": [
          "permanent_red",
          "caution",
          "active_alert",
          "forecast_alert",
          "none"
        ],
        "title": "ZoneClass",
        "description": "Hazard zone classifications (PRD \u00a76.3)."
      }
    }
  }
}